tests/wpt/tests/css/css-masking/animations/clip-path-interpolation-xywh-rect.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>clip-path-interpolation for basic-shape-rect, xywh and rect</title>
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-clip-path">
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#funcdef-basic-shape-xywh">
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#funcdef-basic-shape-rect">
<meta name="assert" content="clip-path supports animation for xywh() and rect()">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<style>
.parent {
  clip-path: inset(10px);
  padding: 10px;
}

.target {
  display: inline-block;
  width: 100px;
  height: 100px;
  background-color: black;
  clip-path: inset(10px);
}

.expected {
  background-color: green;
}
</style>
<body>
<script>
test_no_interpolation({
  property: 'clip-path',
  from: 'initial',
  to: 'xywh(10px 10px 100px 100px)',
});

test_no_interpolation({
  property: 'clip-path',
  from: 'unset',
  to: 'xywh(10px 10px 100px 100px)',
});

test_no_interpolation({
  property: 'clip-path',
  from: 'initial',
  to: 'rect(10px 100px 100px 10px)',
});

test_no_interpolation({
  property: 'clip-path',
  from: 'unset',
  to: 'rect(10px 100px 100px 10px)',
});

// Note: it's fine to use "rect()"/"xywh()" in "expect" because we always
// convert it into `inset()` when we compare it with the actual animation value.
test_interpolation({
  property: 'clip-path',
  from: 'xywh(5px 5px 150% 150%)',
  to: 'xywh(10px 10px 100% 100%)'
}, [
  {at: -1, expect: 'xywh(0px 0px 200% 200%)'},
  {at: 0, expect: 'xywh(5px 5px 150% 150%)'},
  {at: 0.125, expect: 'xywh(5.63px 5.63px 143.75% 143.75%)'},
  {at: 0.875, expect: 'xywh(9.38px 9.38px 106.25% 106.25%)'},
  {at: 1, expect: 'xywh(10px 10px 100% 100%)'},
  {at: 2, expect: 'xywh(15px 15px 50% 50%)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'rect(10px 100px 50% 10%)',
  to: 'rect(50px 200px 90% 50%)'
}, [
  {at: -1, expect: 'rect(-30px 0px 10% -30%)'},
  {at: 0, expect: 'rect(10px 100px 50% 10%)'},
  {at: 0.125, expect: 'rect(15px 112.5px 55% 15%)'},
  {at: 0.875, expect: 'rect(45px 187.5px 85% 45%)'},
  {at: 1, expect: 'rect(50px 200px 90% 50%)'},
  {at: 2, expect: 'rect(90px 300px 130% 90%)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'rect(auto auto auto auto)',
  to: 'rect(80% 20% 20% 80%)'
}, [
  {at: -1, expect: 'inset(-80%)'},
  {at: 0, expect: 'inset(0%)'},
  {at: 0.125, expect: 'inset(10%)'},
  {at: 0.875, expect: 'inset(70%)'},
  {at: 1, expect: 'inset(80%)'},
  {at: 2, expect: 'inset(160%)'},
]);

// All <basic-shape-rect>s compute to the equivalent inset() function, so they
// interpolatable.
// https://drafts.csswg.org/css-shapes-1/#basic-shape-computed-values
test_interpolation({
  property: 'clip-path',
  from: neutralKeyframe,
  // inset(50px calc(80% - 70px) calc(0% - 50px) 20%).
  to: 'xywh(20% 50px 70px 100%)',
}, [
  {at: -1, expect: 'inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)'},
  {at: 0, expect: 'inset(10px calc(0% + 10px) calc(0% + 10px))'},
  {at: 0.125, expect: 'inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))'},
  {at: 0.875, expect: 'inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))'},
  {at: 1, expect: 'inset(50px calc(80% - 70px) calc(0% - 50px) 20%)'},
  {at: 2, expect: 'inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))'},
]);

test_interpolation({
  property: 'clip-path',
  from: neutralKeyframe,
  // inset(50px calc(100% - 70px) 20% 20%).
  to: 'rect(50px 70px 80% 20%)',
}, [
  {at: -1, expect: 'inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)'},
  {at: 0, expect: 'inset(10px calc(0% + 10px) calc(0% + 10px))'},
  {at: 0.125, expect: 'inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))'},
  {at: 0.875, expect: 'inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))'},
  {at: 1, expect: 'inset(50px calc(100% - 70px) 20% 20%)'},
  {at: 2, expect: 'inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'inherit',
  to: 'xywh(20% 50px 70px 100%)',
}, [
  {at: -1, expect: 'inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)'},
  {at: 0, expect: 'inset(10px calc(0% + 10px) calc(0% + 10px))'},
  {at: 0.125, expect: 'inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))'},
  {at: 0.875, expect: 'inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))'},
  {at: 1, expect: 'inset(50px calc(80% - 70px) calc(0% - 50px) 20%)'},
  {at: 2, expect: 'inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'inherit',
  to: 'rect(50px 70px 80% 20%)',
}, [
  {at: -1, expect: 'inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)'},
  {at: 0, expect: 'inset(10px calc(0% + 10px) calc(0% + 10px))'},
  {at: 0.125, expect: 'inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))'},
  {at: 0.875, expect: 'inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))'},
  {at: 1, expect: 'inset(50px calc(100% - 70px) 20% 20%)'},
  {at: 2, expect: 'inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'xywh(0px 10% 100px 40% round 20px)',
  to: 'rect(20% 50% 200px 20px)',
}, [
  {at: -1, expect: 'inset(0% calc(150% - 200px) calc(0% + 200px) -20px round 40px'},
  {at: 0, expect: 'inset(10% calc(100% - 100px) 50% 0px round 20px)'},
  {at: 0.125, expect: 'inset(11.25% calc(93.75% - 87.5px) calc(56.25% - 25px) 2.5px round 17.5px)'},
  {at: 0.875, expect: 'inset(18.75% calc(56.25% - 12.5px) calc(93.75% - 175px) 17.5px round 2.5px)'},
  {at: 1, expect: 'inset(20% 50% calc(100% - 200px) 20px)'},
  {at: 2, expect: 'inset(30% calc(0% + 100px) calc(150% - 400px) 40px)'},
]);

</script>
</body>
