third_party/blink/web_tests/external/wpt/css/css-masking/animations/clip-path-interpolation-xywh-rect-expected.txt

This is a testharness.js-based test.
Found 304 tests; 198 PASS, 106 FAIL, 0 TIMEOUT, 0 NOTRUN.
[PASS] CSS Transitions: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (-0.3) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.3) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.6) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (1) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (1.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (-0.3) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.3) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.6) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (1) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (1.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (-0.3) should be [initial]
[PASS] CSS Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0) should be [initial]
[PASS] CSS Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.3) should be [initial]
[PASS] CSS Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.6) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (1) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (1.5) should be [xywh(10px 10px 100px 100px)]
[PASS] Web Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (-0.3) should be [initial]
[PASS] Web Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0) should be [initial]
[PASS] Web Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.3) should be [initial]
[PASS] Web Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.5) should be [xywh(10px 10px 100px 100px)]
[PASS] Web Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (0.6) should be [xywh(10px 10px 100px 100px)]
[PASS] Web Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (1) should be [xywh(10px 10px 100px 100px)]
[PASS] Web Animations: property <clip-path> from [initial] to [xywh(10px 10px 100px 100px)] at (1.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (-0.3) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.3) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.6) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (1) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (1.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (-0.3) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.3) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.6) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (1) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (1.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (-0.3) should be [unset]
[PASS] CSS Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0) should be [unset]
[PASS] CSS Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.3) should be [unset]
[PASS] CSS Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.6) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (1) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (1.5) should be [xywh(10px 10px 100px 100px)]
[PASS] Web Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (-0.3) should be [unset]
[PASS] Web Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0) should be [unset]
[PASS] Web Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.3) should be [unset]
[PASS] Web Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.5) should be [xywh(10px 10px 100px 100px)]
[PASS] Web Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (0.6) should be [xywh(10px 10px 100px 100px)]
[PASS] Web Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (1) should be [xywh(10px 10px 100px 100px)]
[PASS] Web Animations: property <clip-path> from [unset] to [xywh(10px 10px 100px 100px)] at (1.5) should be [xywh(10px 10px 100px 100px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (-0.3) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.3) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.6) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (1) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (1.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (-0.3) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.3) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.6) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (1) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (1.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (-0.3) should be [initial]
[PASS] CSS Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0) should be [initial]
[PASS] CSS Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.3) should be [initial]
[PASS] CSS Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.6) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (1) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (1.5) should be [rect(10px 100px 100px 10px)]
[PASS] Web Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (-0.3) should be [initial]
[PASS] Web Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0) should be [initial]
[PASS] Web Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.3) should be [initial]
[PASS] Web Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.5) should be [rect(10px 100px 100px 10px)]
[PASS] Web Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (0.6) should be [rect(10px 100px 100px 10px)]
[PASS] Web Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (1) should be [rect(10px 100px 100px 10px)]
[PASS] Web Animations: property <clip-path> from [initial] to [rect(10px 100px 100px 10px)] at (1.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (-0.3) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.3) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.6) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (1) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (1.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (-0.3) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.3) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.6) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (1) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (1.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (-0.3) should be [unset]
[PASS] CSS Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0) should be [unset]
[PASS] CSS Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.3) should be [unset]
[PASS] CSS Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.6) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (1) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (1.5) should be [rect(10px 100px 100px 10px)]
[PASS] Web Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (-0.3) should be [unset]
[PASS] Web Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0) should be [unset]
[PASS] Web Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.3) should be [unset]
[PASS] Web Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.5) should be [rect(10px 100px 100px 10px)]
[PASS] Web Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (0.6) should be [rect(10px 100px 100px 10px)]
[PASS] Web Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (1) should be [rect(10px 100px 100px 10px)]
[PASS] Web Animations: property <clip-path> from [unset] to [rect(10px 100px 100px 10px)] at (1.5) should be [rect(10px 100px 100px 10px)]
[PASS] CSS Transitions: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (-1) should be [xywh(0px 0px 200% 200%)]
[PASS] CSS Transitions: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0) should be [xywh(5px 5px 150% 150%)]
[PASS] CSS Transitions: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0.125) should be [xywh(5.63px 5.63px 143.75% 143.75%)]
[PASS] CSS Transitions: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0.875) should be [xywh(9.38px 9.38px 106.25% 106.25%)]
[PASS] CSS Transitions: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (1) should be [xywh(10px 10px 100% 100%)]
[PASS] CSS Transitions: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (2) should be [xywh(15px 15px 50% 50%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (-1) should be [xywh(0px 0px 200% 200%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0) should be [xywh(5px 5px 150% 150%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0.125) should be [xywh(5.63px 5.63px 143.75% 143.75%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0.875) should be [xywh(9.38px 9.38px 106.25% 106.25%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (1) should be [xywh(10px 10px 100% 100%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (2) should be [xywh(15px 15px 50% 50%)]
[PASS] CSS Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (-1) should be [xywh(0px 0px 200% 200%)]
[PASS] CSS Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0) should be [xywh(5px 5px 150% 150%)]
[PASS] CSS Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0.125) should be [xywh(5.63px 5.63px 143.75% 143.75%)]
[PASS] CSS Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0.875) should be [xywh(9.38px 9.38px 106.25% 106.25%)]
[PASS] CSS Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (1) should be [xywh(10px 10px 100% 100%)]
[PASS] CSS Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (2) should be [xywh(15px 15px 50% 50%)]
[PASS] Web Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (-1) should be [xywh(0px 0px 200% 200%)]
[PASS] Web Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0) should be [xywh(5px 5px 150% 150%)]
[PASS] Web Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0.125) should be [xywh(5.63px 5.63px 143.75% 143.75%)]
[PASS] Web Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (0.875) should be [xywh(9.38px 9.38px 106.25% 106.25%)]
[PASS] Web Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (1) should be [xywh(10px 10px 100% 100%)]
[PASS] Web Animations: property <clip-path> from [xywh(5px 5px 150% 150%)] to [xywh(10px 10px 100% 100%)] at (2) should be [xywh(15px 15px 50% 50%)]
[PASS] CSS Transitions: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (-1) should be [rect(-30px 0px 10% -30%)]
[PASS] CSS Transitions: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0) should be [rect(10px 100px 50% 10%)]
[PASS] CSS Transitions: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0.125) should be [rect(15px 112.5px 55% 15%)]
[PASS] CSS Transitions: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0.875) should be [rect(45px 187.5px 85% 45%)]
[PASS] CSS Transitions: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (1) should be [rect(50px 200px 90% 50%)]
[PASS] CSS Transitions: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (2) should be [rect(90px 300px 130% 90%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (-1) should be [rect(-30px 0px 10% -30%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0) should be [rect(10px 100px 50% 10%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0.125) should be [rect(15px 112.5px 55% 15%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0.875) should be [rect(45px 187.5px 85% 45%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (1) should be [rect(50px 200px 90% 50%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (2) should be [rect(90px 300px 130% 90%)]
[PASS] CSS Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (-1) should be [rect(-30px 0px 10% -30%)]
[PASS] CSS Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0) should be [rect(10px 100px 50% 10%)]
[PASS] CSS Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0.125) should be [rect(15px 112.5px 55% 15%)]
[PASS] CSS Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0.875) should be [rect(45px 187.5px 85% 45%)]
[PASS] CSS Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (1) should be [rect(50px 200px 90% 50%)]
[PASS] CSS Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (2) should be [rect(90px 300px 130% 90%)]
[PASS] Web Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (-1) should be [rect(-30px 0px 10% -30%)]
[PASS] Web Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0) should be [rect(10px 100px 50% 10%)]
[PASS] Web Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0.125) should be [rect(15px 112.5px 55% 15%)]
[PASS] Web Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (0.875) should be [rect(45px 187.5px 85% 45%)]
[PASS] Web Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (1) should be [rect(50px 200px 90% 50%)]
[PASS] Web Animations: property <clip-path> from [rect(10px 100px 50% 10%)] to [rect(50px 200px 90% 50%)] at (2) should be [rect(90px 300px 130% 90%)]
[PASS] CSS Transitions: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (-1) should be [inset(-80%)]
[PASS] CSS Transitions: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0) should be [inset(0%)]
[PASS] CSS Transitions: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0.125) should be [inset(10%)]
[PASS] CSS Transitions: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0.875) should be [inset(70%)]
[PASS] CSS Transitions: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (1) should be [inset(80%)]
[PASS] CSS Transitions: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (2) should be [inset(160%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (-1) should be [inset(-80%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0) should be [inset(0%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0.125) should be [inset(10%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0.875) should be [inset(70%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (1) should be [inset(80%)]
[PASS] CSS Transitions with transition: all: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (2) should be [inset(160%)]
[PASS] CSS Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (-1) should be [inset(-80%)]
[PASS] CSS Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0) should be [inset(0%)]
[PASS] CSS Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0.125) should be [inset(10%)]
[PASS] CSS Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0.875) should be [inset(70%)]
[PASS] CSS Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (1) should be [inset(80%)]
[PASS] CSS Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (2) should be [inset(160%)]
[PASS] Web Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (-1) should be [inset(-80%)]
[PASS] Web Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0) should be [inset(0%)]
[PASS] Web Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0.125) should be [inset(10%)]
[PASS] Web Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (0.875) should be [inset(70%)]
[PASS] Web Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (1) should be [inset(80%)]
[PASS] Web Animations: property <clip-path> from [rect(auto auto auto auto)] to [rect(80% 20% 20% 80%)] at (2) should be [inset(160%)]
[FAIL] CSS Transitions: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (-1) should be [inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 80 % + 90px ) calc ( 0 % + 70px ) calc ( - 20 % + 20px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0.125) should be [inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 10 % calc ( 0 % + 2.5px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0.875) should be [inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 70 % - 60px ) calc ( 0 % - 42.5px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (1) should be [inset(50px calc(80% - 70px) calc(0% - 50px) 20%)]
  assert_equals: expected "inset ( 50px calc ( 80 % - 70px ) calc ( 0 % - 50px ) 20 % ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (2) should be [inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 160 % - 150px ) calc ( 0 % - 110px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (-1) should be [inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 80 % + 90px ) calc ( 0 % + 70px ) calc ( - 20 % + 20px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0.125) should be [inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 10 % calc ( 0 % + 2.5px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0.875) should be [inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 70 % - 60px ) calc ( 0 % - 42.5px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (1) should be [inset(50px calc(80% - 70px) calc(0% - 50px) 20%)]
  assert_equals: expected "inset ( 50px calc ( 80 % - 70px ) calc ( 0 % - 50px ) 20 % ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (2) should be [inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 160 % - 150px ) calc ( 0 % - 110px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (-1) should be [inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 80 % + 90px ) calc ( 0 % + 70px ) calc ( - 20 % + 20px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0.125) should be [inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 10 % calc ( 0 % + 2.5px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0.875) should be [inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 70 % - 60px ) calc ( 0 % - 42.5px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (1) should be [inset(50px calc(80% - 70px) calc(0% - 50px) 20%)]
  assert_equals: expected "inset ( 50px calc ( 80 % - 70px ) calc ( 0 % - 50px ) 20 % ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (2) should be [inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 160 % - 150px ) calc ( 0 % - 110px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] Web Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (-1) should be [inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 80 % + 90px ) calc ( 0 % + 70px ) calc ( - 20 % + 20px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0.125) should be [inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 10 % calc ( 0 % + 2.5px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (0.875) should be [inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 70 % - 60px ) calc ( 0 % - 42.5px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] Web Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (1) should be [inset(50px calc(80% - 70px) calc(0% - 50px) 20%)]
  assert_equals: expected "inset ( 50px calc ( 80 % - 70px ) calc ( 0 % - 50px ) 20 % ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] Web Animations: property <clip-path> from neutral to [xywh(20% 50px 70px 100%)] at (2) should be [inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 160 % - 150px ) calc ( 0 % - 110px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (-1) should be [inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 100 % + 90px ) calc ( - 20 % + 20px ) calc ( - 20 % + 20px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0.125) should be [inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 12.5 % calc ( 2.5 % + 8.75px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0.875) should be [inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 87.5 % - 60px ) calc ( 17.5 % + 1.25px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[PASS] CSS Transitions: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (1) should be [inset(50px calc(100% - 70px) 20% 20%)]
[FAIL] CSS Transitions: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (2) should be [inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 200 % - 150px ) calc ( 40 % - 10px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (-1) should be [inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 100 % + 90px ) calc ( - 20 % + 20px ) calc ( - 20 % + 20px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0.125) should be [inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 12.5 % calc ( 2.5 % + 8.75px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0.875) should be [inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 87.5 % - 60px ) calc ( 17.5 % + 1.25px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[PASS] CSS Transitions with transition: all: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (1) should be [inset(50px calc(100% - 70px) 20% 20%)]
[FAIL] CSS Transitions with transition: all: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (2) should be [inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 200 % - 150px ) calc ( 40 % - 10px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (-1) should be [inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 100 % + 90px ) calc ( - 20 % + 20px ) calc ( - 20 % + 20px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0.125) should be [inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 12.5 % calc ( 2.5 % + 8.75px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0.875) should be [inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 87.5 % - 60px ) calc ( 17.5 % + 1.25px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[PASS] CSS Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (1) should be [inset(50px calc(100% - 70px) 20% 20%)]
[FAIL] CSS Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (2) should be [inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 200 % - 150px ) calc ( 40 % - 10px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] Web Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (-1) should be [inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 100 % + 90px ) calc ( - 20 % + 20px ) calc ( - 20 % + 20px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0.125) should be [inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 12.5 % calc ( 2.5 % + 8.75px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (0.875) should be [inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 87.5 % - 60px ) calc ( 17.5 % + 1.25px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[PASS] Web Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (1) should be [inset(50px calc(100% - 70px) 20% 20%)]
[FAIL] Web Animations: property <clip-path> from neutral to [rect(50px 70px 80% 20%)] at (2) should be [inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 200 % - 150px ) calc ( 40 % - 10px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (-1) should be [inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 80 % + 90px ) calc ( 0 % + 70px ) calc ( - 20 % + 20px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0.125) should be [inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 10 % calc ( 0 % + 2.5px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0.875) should be [inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 70 % - 60px ) calc ( 0 % - 42.5px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (1) should be [inset(50px calc(80% - 70px) calc(0% - 50px) 20%)]
  assert_equals: expected "inset ( 50px calc ( 80 % - 70px ) calc ( 0 % - 50px ) 20 % ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (2) should be [inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 160 % - 150px ) calc ( 0 % - 110px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (-1) should be [inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 80 % + 90px ) calc ( 0 % + 70px ) calc ( - 20 % + 20px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0.125) should be [inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 10 % calc ( 0 % + 2.5px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0.875) should be [inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 70 % - 60px ) calc ( 0 % - 42.5px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (1) should be [inset(50px calc(80% - 70px) calc(0% - 50px) 20%)]
  assert_equals: expected "inset ( 50px calc ( 80 % - 70px ) calc ( 0 % - 50px ) 20 % ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (2) should be [inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 160 % - 150px ) calc ( 0 % - 110px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (-1) should be [inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 80 % + 90px ) calc ( 0 % + 70px ) calc ( - 20 % + 20px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0.125) should be [inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 10 % calc ( 0 % + 2.5px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0.875) should be [inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 70 % - 60px ) calc ( 0 % - 42.5px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (1) should be [inset(50px calc(80% - 70px) calc(0% - 50px) 20%)]
  assert_equals: expected "inset ( 50px calc ( 80 % - 70px ) calc ( 0 % - 50px ) 20 % ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (2) should be [inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 160 % - 150px ) calc ( 0 % - 110px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (-1) should be [inset(-30px calc(-80% + 90px) calc(0% + 70px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 80 % + 90px ) calc ( 0 % + 70px ) calc ( - 20 % + 20px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0.125) should be [inset(15px 10% calc(0% + 2.5px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 10 % calc ( 0 % + 2.5px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (0.875) should be [inset(45px calc(70% - 60px) calc(0% - 42.5px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 70 % - 60px ) calc ( 0 % - 42.5px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (1) should be [inset(50px calc(80% - 70px) calc(0% - 50px) 20%)]
  assert_equals: expected "inset ( 50px calc ( 80 % - 70px ) calc ( 0 % - 50px ) 20 % ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [xywh(20% 50px 70px 100%)] at (2) should be [inset(90px calc(160% - 150px) calc(0% - 110px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 160 % - 150px ) calc ( 0 % - 110px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 80 % - 70px ) - 50px 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (-1) should be [inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 100 % + 90px ) calc ( - 20 % + 20px ) calc ( - 20 % + 20px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0.125) should be [inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 12.5 % calc ( 2.5 % + 8.75px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0.875) should be [inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 87.5 % - 60px ) calc ( 17.5 % + 1.25px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[PASS] CSS Transitions: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (1) should be [inset(50px calc(100% - 70px) 20% 20%)]
[FAIL] CSS Transitions: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (2) should be [inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 200 % - 150px ) calc ( 40 % - 10px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (-1) should be [inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 100 % + 90px ) calc ( - 20 % + 20px ) calc ( - 20 % + 20px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0.125) should be [inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 12.5 % calc ( 2.5 % + 8.75px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0.875) should be [inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 87.5 % - 60px ) calc ( 17.5 % + 1.25px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[PASS] CSS Transitions with transition: all: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (1) should be [inset(50px calc(100% - 70px) 20% 20%)]
[FAIL] CSS Transitions with transition: all: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (2) should be [inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 200 % - 150px ) calc ( 40 % - 10px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (-1) should be [inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 100 % + 90px ) calc ( - 20 % + 20px ) calc ( - 20 % + 20px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0.125) should be [inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 12.5 % calc ( 2.5 % + 8.75px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 10px ) "
[FAIL] CSS Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0.875) should be [inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 87.5 % - 60px ) calc ( 17.5 % + 1.25px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[PASS] CSS Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (1) should be [inset(50px calc(100% - 70px) 20% 20%)]
[FAIL] CSS Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (2) should be [inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 200 % - 150px ) calc ( 40 % - 10px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (-1) should be [inset(-30px calc(-100% + 90px) calc(-20% + 20px) calc(-20% + 20px)]
  assert_equals: expected "inset ( - 30px calc ( - 100 % + 90px ) calc ( - 20 % + 20px ) calc ( - 20 % + 20px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0) should be [inset(10px calc(0% + 10px) calc(0% + 10px))]
  assert_equals: expected "inset ( 10px calc ( 0 % + 10px ) calc ( 0 % + 10px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0.125) should be [inset(15px 12.5% calc(2.5% + 8.75px) calc(2.5% + 8.75px))]
  assert_equals: expected "inset ( 15px 12.5 % calc ( 2.5 % + 8.75px ) calc ( 2.5 % + 8.75px ) ) " but got "inset ( 10px ) "
[FAIL] Web Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (0.875) should be [inset(45px calc(87.5% - 60px) calc(17.5% + 1.25px) calc(17.5% + 1.25px))]
  assert_equals: expected "inset ( 45px calc ( 87.5 % - 60px ) calc ( 17.5 % + 1.25px ) calc ( 17.5 % + 1.25px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[PASS] Web Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (1) should be [inset(50px calc(100% - 70px) 20% 20%)]
[FAIL] Web Animations: property <clip-path> from [inherit] to [rect(50px 70px 80% 20%)] at (2) should be [inset(90px calc(200% - 150px) calc(40% - 10px) calc(40% - 10px))]
  assert_equals: expected "inset ( 90px calc ( 200 % - 150px ) calc ( 40 % - 10px ) calc ( 40 % - 10px ) ) " but got "inset ( 50px calc ( 100 % - 70px ) 20 % 20 % ) "
[FAIL] CSS Transitions: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (-1) should be [inset(0% calc(150% - 200px) calc(0% + 200px) -20px round 40px]
  assert_equals: expected "inset ( 0 % calc ( 150 % - 200px ) calc ( 0 % + 200px ) - 20px round 40px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[FAIL] CSS Transitions: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0) should be [inset(10% calc(100% - 100px) 50% 0px round 20px)]
  assert_equals: expected "inset ( 10 % calc ( 100 % - 100px ) 50 % 0px round 20px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[FAIL] CSS Transitions: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0.125) should be [inset(11.25% calc(93.75% - 87.5px) calc(56.25% - 25px) 2.5px round 17.5px)]
  assert_equals: expected "inset ( 11.25 % calc ( 93.75 % - 87.5px ) calc ( 56.25 % - 25px ) 2.5px round 17.5px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[FAIL] CSS Transitions: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0.875) should be [inset(18.75% calc(56.25% - 12.5px) calc(93.75% - 175px) 17.5px round 2.5px)]
  assert_equals: expected "inset ( 18.75 % calc ( 56.25 % - 12.5px ) calc ( 93.75 % - 175px ) 17.5px round 2.5px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[PASS] CSS Transitions: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (1) should be [inset(20% 50% calc(100% - 200px) 20px)]
[FAIL] CSS Transitions: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (2) should be [inset(30% calc(0% + 100px) calc(150% - 400px) 40px)]
  assert_equals: expected "inset ( 30 % calc ( 0 % + 100px ) calc ( 150 % - 400px ) 40px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (-1) should be [inset(0% calc(150% - 200px) calc(0% + 200px) -20px round 40px]
  assert_equals: expected "inset ( 0 % calc ( 150 % - 200px ) calc ( 0 % + 200px ) - 20px round 40px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0) should be [inset(10% calc(100% - 100px) 50% 0px round 20px)]
  assert_equals: expected "inset ( 10 % calc ( 100 % - 100px ) 50 % 0px round 20px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0.125) should be [inset(11.25% calc(93.75% - 87.5px) calc(56.25% - 25px) 2.5px round 17.5px)]
  assert_equals: expected "inset ( 11.25 % calc ( 93.75 % - 87.5px ) calc ( 56.25 % - 25px ) 2.5px round 17.5px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[FAIL] CSS Transitions with transition: all: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0.875) should be [inset(18.75% calc(56.25% - 12.5px) calc(93.75% - 175px) 17.5px round 2.5px)]
  assert_equals: expected "inset ( 18.75 % calc ( 56.25 % - 12.5px ) calc ( 93.75 % - 175px ) 17.5px round 2.5px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[PASS] CSS Transitions with transition: all: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (1) should be [inset(20% 50% calc(100% - 200px) 20px)]
[FAIL] CSS Transitions with transition: all: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (2) should be [inset(30% calc(0% + 100px) calc(150% - 400px) 40px)]
  assert_equals: expected "inset ( 30 % calc ( 0 % + 100px ) calc ( 150 % - 400px ) 40px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[FAIL] CSS Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (-1) should be [inset(0% calc(150% - 200px) calc(0% + 200px) -20px round 40px]
  assert_equals: expected "inset ( 0 % calc ( 150 % - 200px ) calc ( 0 % + 200px ) - 20px round 40px ) " but got "inset ( 10 % calc ( 100 % - 100px ) 50 % 0px round 20px ) "
[PASS] CSS Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0) should be [inset(10% calc(100% - 100px) 50% 0px round 20px)]
[FAIL] CSS Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0.125) should be [inset(11.25% calc(93.75% - 87.5px) calc(56.25% - 25px) 2.5px round 17.5px)]
  assert_equals: expected "inset ( 11.25 % calc ( 93.75 % - 87.5px ) calc ( 56.25 % - 25px ) 2.5px round 17.5px ) " but got "inset ( 10 % calc ( 100 % - 100px ) 50 % 0px round 20px ) "
[FAIL] CSS Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0.875) should be [inset(18.75% calc(56.25% - 12.5px) calc(93.75% - 175px) 17.5px round 2.5px)]
  assert_equals: expected "inset ( 18.75 % calc ( 56.25 % - 12.5px ) calc ( 93.75 % - 175px ) 17.5px round 2.5px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[PASS] CSS Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (1) should be [inset(20% 50% calc(100% - 200px) 20px)]
[FAIL] CSS Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (2) should be [inset(30% calc(0% + 100px) calc(150% - 400px) 40px)]
  assert_equals: expected "inset ( 30 % calc ( 0 % + 100px ) calc ( 150 % - 400px ) 40px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[FAIL] Web Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (-1) should be [inset(0% calc(150% - 200px) calc(0% + 200px) -20px round 40px]
  assert_equals: expected "inset ( 0 % calc ( 150 % - 200px ) calc ( 0 % + 200px ) - 20px round 40px ) " but got "inset ( 10 % calc ( 100 % - 100px ) 50 % 0px round 20px ) "
[PASS] Web Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0) should be [inset(10% calc(100% - 100px) 50% 0px round 20px)]
[FAIL] Web Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0.125) should be [inset(11.25% calc(93.75% - 87.5px) calc(56.25% - 25px) 2.5px round 17.5px)]
  assert_equals: expected "inset ( 11.25 % calc ( 93.75 % - 87.5px ) calc ( 56.25 % - 25px ) 2.5px round 17.5px ) " but got "inset ( 10 % calc ( 100 % - 100px ) 50 % 0px round 20px ) "
[FAIL] Web Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (0.875) should be [inset(18.75% calc(56.25% - 12.5px) calc(93.75% - 175px) 17.5px round 2.5px)]
  assert_equals: expected "inset ( 18.75 % calc ( 56.25 % - 12.5px ) calc ( 93.75 % - 175px ) 17.5px round 2.5px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
[PASS] Web Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (1) should be [inset(20% 50% calc(100% - 200px) 20px)]
[FAIL] Web Animations: property <clip-path> from [xywh(0px 10% 100px 40% round 20px)] to [rect(20% 50% 200px 20px)] at (2) should be [inset(30% calc(0% + 100px) calc(150% - 400px) 40px)]
  assert_equals: expected "inset ( 30 % calc ( 0 % + 100px ) calc ( 150 % - 400px ) 40px ) " but got "inset ( 20 % 50 % calc ( 100 % - 200px ) 20px ) "
Harness: the test ran to completion.
